+++ test/utils/testutils.ts
@@ -20,7 +20,7 @@ import {
   parseEther
 } from 'ethers/lib/utils'
 import { ethers } from 'hardhat'
-import { debugTracers } from './_debugTx'
+import { debugTracers } from './debugTx'
 import { UserOperation } from './UserOperation'
 
 export async function createAccountFromFactory (
